Fixes suggested by running 'splint -weak *.c' on code.  Specifically making variable casts explicit and eliminating use of sprintf in favor of snprintf. No actual bugs were uncovered or fixed in this process.
git-svn-id: https://latex2rtf.svn.sourceforge.net/svnroot/latex2rtf/trunk@331 e562ef82-0334-0410-8083-dbe22b1e9a75

+++ definitions.h
@@ -1,10 +1,10 @@
-int     maybeDefinition(char * s, int n);
+int     maybeDefinition(char * s, size_t n);
 int     existsDefinition(char * s);
 void	newDefinition(char *name, char *def, int params);
 void	renewDefinition(char * name, char * def, int params);
 char *	expandDefinition(int thedef);
 
-int		maybeEnvironment(char * s, int n);
+int		maybeEnvironment(char * s, size_t n);
 int		existsEnvironment(char * s);
 void	newEnvironment(char *name, char *begdef, char *enddef, int params);
 void	renewEnvironment(char *name, char *begdef, char *enddef, int params);
